crypto_apis/crypto_apis/app/requirements.txt
pyodbc==4.0.32
Django==3.2.8
djangorestframework==3.12.3
requests==2.25.0
python-decouple==3.4
mssql-django==1.0
pytest==6.1.2
pytest-django==4.1.0